Call API with proper input argumnets to add message to provider
Android uses addMessageToUri in telephony to add messages from
third party applications, but while calling the function it is
not passing last argument type(threadId) properly, so instead
 of actual API it is calling MSIM overloaded API.

CRs-Fixed: 590125

+++ src/java/com/android/internal/telephony/SMSDispatcher.java
@@ -1092,7 +1092,7 @@ public abstract class SMSDispatcher extends Handler {
                         mTimestamp /*date*/,
                         true /*read*/,
                         deliveryReport /*deliveryReport*/,
-                        0 /*threadId*/);
+                        (long) 0 /*threadId*/);
             }
         }
